Mock Xtream server, streaming JSON reader, discovery report
Mock server (stdlib Python): player_api.php, xmltv.php, endless TS and a
sliding-window HLS playlist over FFmpeg-generated colour-bar media. 5,181
channels in 213 categories across 44 countries, mixed prefix styles, locals for
28 US markets. Serves a 187 MB XMLTV with 542,046 programmes, gzip on request.

Every generated channel carries ground truth, so the discovery tool scores the
filter rather than eyeballing it: precision 1.0000, recall 1.0000, zero
disagreements. The category short-circuit avoids downloading 4,629 of 5,181
channels.

JsonReader is a pull parser in core, used by both the import pipeline and the
discovery tool, so the report exercises the shipping filter.

Fixes a real bug found by the ground-truth scoring: stream_id was parsed as Int,
so ids past Int.MAX_VALUE clamped onto a single value and collapsed half the
catalogue onto one key. Stream ids are Long throughout, and nextInt now returns
0 out of range rather than clamping silently.

+++ core/src/main/kotlin/com/retroguide/core/model/Models.kt
@@ -42,7 +42,7 @@ data class RawCategory(
 
 /** A live stream exactly as `get_live_streams` returned it, before any filtering. */
 data class RawChannel(
-    val streamId: Int,
+    val streamId: Long,
     val name: String,
     val categoryId: String?,
     val epgChannelId: String? = null,
@@ -106,7 +106,7 @@ data class FilterDecision(
 
 /** A channel that survived the filter, ready to be written to the database. */
 data class KeptChannel(
-    val streamId: Int,
+    val streamId: Long,
     val originalName: String,
     val displayName: String,
     val country: Country,
